Auth & debug

+++ backend/routes/todos.js
@@ -9,17 +9,15 @@ router.get("/", auth, async(req, res) => {
     try {
         const todos = await Todo.find()
         .sort({ date: -1 })
-
         const filteredTodos = todos.filter( todo => todo.uid === req.user._id)
-
-        res.send(todos)
+        res.send(filteredTodos);
     } catch(error) {
         res.status(500).send(error.message)
         console.log(error)
     }
 }) 
 
-router.post("/", async(req, res) => {
+router.post("/", auth, async(req, res) => {
     const schema = Joi.object({
         name: Joi.string().min(3).max(200).required(),
         author: Joi.string().min(3).max(30),
@@ -51,7 +49,7 @@ router.post("/", async(req, res) => {
     } 
 })
 //Edit the body of a document
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
     const schema = Joi.object({
         name: Joi.string().min(3).max(200).required(),
         author: Joi.string().min(3).max(30),
@@ -68,6 +66,9 @@ router.put("/:id", async (req, res) => {
     const todo = await Todo.findById(req.params.id)
 
     if(!todo) return res.status(404).send("Todo not found...")
+
+    if(todo.uid !== req.user._id) 
+        return res.status(401).send("Todo update failed. Not authorized...")
 
     const { name, author, isComplete, date, uid } = req.body
 
@@ -87,12 +88,15 @@ router.put("/:id", async (req, res) => {
     }
 })
 
-router.patch("/:id", async (req, res) => {
+router.patch("/:id", auth, async (req, res) => {
     try {
     
         const todo = await Todo.findById(req.params.id)
 
         if(!todo) return res.status(404).send("Todo not found...")
+
+        if(todo.uid !== req.user._id) 
+            return res.status(401).send("Todo check/uncheck failed. Not authorized...")
 
         const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, {
         isComplete: !todo.isComplete
@@ -104,11 +108,14 @@ router.patch("/:id", async (req, res) => {
     }
 })
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
     try {
         const todo = await Todo.findById(req.params.id)
 
         if(!todo) return res.status(404).send("Todo not found...")
+
+        if(todo.uid !== req.user._id) 
+            return res.status(401).send("Todo deletion failed. Not authorized...")
 
         const deletedTodo = await Todo.findByIdAndDelete(req.params.id)
